feat: 이전 버전 파일 제거 및 임시 파일 이름 변경 (#139)
