+++ Projects/Features/TopicFeature/Sources/ViewController/Generate/TopicGenerateViewController.swift
@@ -1,9 +1,9 @@
 //
-//  TopicGenerateViewController.swift
+//  TopicGenerateViewControllerTest.swift
 //  TopicFeature
 //
-//  Created by 박소윤 on 2023/12/25.
-//  Copyright © 2023 AB. All rights reserved.
+//  Created by 박소윤 on 2024/01/20.
+//  Copyright © 2024 AB. All rights reserved.
 //
 
 import Foundation
@@ -12,37 +12,74 @@ import ABKit
 import TopicFeatureInterface
 import FeatureDependency
 import Domain
-import Combine
-import Core
-import PhotosUI
 
-/*
-final class TopicGenerateViewController: BaseViewController<TopicGenerateHeaderView, TopicGenerateView, DefaultTopicGenerateCoordinator> {
-
+final class TopicGenerateViewController: BaseViewController<TopicGenerateHeaderView, BaseView, DefaultTopicGenerateCoordinator> {
+    
     init(viewModel: any TopicGenerateViewModel) {
         self.viewModel = viewModel
-        super.init(headerView: TopicGenerateHeaderView(), mainView: TopicGenerateView())
+        super.init(headerView: TopicGenerateHeaderView(), mainView: BaseView())
     }
-    
-    private let viewModel: any TopicGenerateViewModel
-    private var inputCell: TopicInputTableViewCell?
-    private var contentInputCell: TopicContentInputTableViewCell?
-    private var imagePickerOption: Choice.Option?
-    private var sideChangeButton: TopicSideChangeButton?
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
     
-    override func initialize() {
+    private var viewModel: any TopicGenerateViewModel
+    private let aSideView: TopicGenerateASideView = TopicGenerateASideView()
+    private let bSideView: TopicGenerateBSideFirstView = TopicGenerateBSideFirstView()
     
-        setTableViewDelegate()
+    private var sideChangeButton: TopicSideChangeButton?
+    private var currentSideView: UIView? {
+        willSet {
+            newValue?.isHidden = false
+        }
+        didSet {
+            oldValue?.isHidden = true
+        }
+    }
+    
+    override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
+        view.endEditing(true)
+    }
+    
+    override func initialize() {
+        
+        setLimitCount()
         setHeaderViewTopicSide()
         addSideChangeTarget()
+        collectionViewDelegate()
         
-        func setTableViewDelegate() {
-            mainView.tableView.delegate = self
-            mainView.tableView.dataSource = self
+        mainView.addSubviews([aSideView, bSideView])
+        [aSideView, bSideView].forEach{ view in
+            view.isHidden = true
+            view.snp.makeConstraints{
+                $0.top.leading.trailing.bottom.equalToSuperview()
+            }
+        }
+        
+        aSideView.optionSwitch.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(optionSwitch)))
+        
+        aSideView.ctaButton.tapPublisher
+            .sink{ [weak self] _ in
+                guard let self = self else { return }
+                self.viewModel.register(.init(
+                    side: .A,
+                    keyword: nil,
+                    title: self.aSideView.titleSection.contentView.textField.text ?? "",
+                    choices: [.init(text: self.aSideView.optionsSection.contentView.aTextField.textField.text, image: nil, option: .A),
+                              .init(text: self.aSideView.optionsSection.contentView.aTextField.textField.text, image: nil, option: .B)],
+                    deadline: nil
+                ))
+            }
+            .store(in: &cancellables)
+        
+        func setLimitCount() {
+            aSideView.titleSection.contentView.limitCount = viewModel.limitCount.title
+            aSideView.optionsSection.contentView.aTextField.limitCount = viewModel.limitCount.textOption
+            aSideView.optionsSection.contentView.bTextField.limitCount = viewModel.limitCount.textOption
+            
+            bSideView.titleSection.contentView.limitCount = viewModel.limitCount.title
+            bSideView.keywordSection.contentView.limitCount = viewModel.limitCount.keyword
         }
         
         func setHeaderViewTopicSide() {
@@ -92,157 +129,154 @@ final class TopicGenerateViewController: BaseViewController<TopicGenerateHeaderV
                 sideChangeButton = nil
             }
         }
+        
+        func collectionViewDelegate() {
+            bSideView.recommendKeyword.collectionView.delegate = self
+            bSideView.recommendKeyword.collectionView.dataSource = self
+        }
+    }
+    
+    @objc private func optionSwitch() {
+        let temp = aSideView.optionsSection.contentView.aTextField.textField.text ?? ""
+        aSideView.optionsSection.contentView.aTextField.update(text: aSideView.optionsSection.contentView.bTextField.textField.text ?? "")
+        aSideView.optionsSection.contentView.bTextField.update(text: temp)
     }
     
     override func bind() {
         
-        NotificationCenter.default
-            .publisher(for: Notification.Name( Topic.Action.showImagePicker.identifier), object: contentInputCell)
-            .receive(on: DispatchQueue.main)
-            .sink{ [weak self] object in
-                
-                guard let self = self else { return }
-                
-                self.imagePickerOption = object.userInfo?[Choice.Option.identifier] as? Choice.Option
-                self.present(imagePicker(), animated: true)
-                
-                func imagePicker() -> PHPickerViewController {
-                    var configuration = PHPickerConfiguration()
-                    configuration.selectionLimit = 1
-                    configuration.filter = .images
-                    let picker = PHPickerViewController(configuration: configuration)
-                    picker.delegate = self
-                    return picker
-                }
-            }
-            .store(in: &cancellables)
-        
-        viewModel.contentValidation
-            .receive(on: DispatchQueue.main)
-            .sink{ [weak self] isValid in
-                self?.mainView.tableView.isScrollEnabled = isValid
-                if isValid {
-                    self?.mainView.tableView.scrollToRow(at: IndexPath(row: 1), at: .top, animated: true)
-                }
-            }
-            .store(in: &cancellables)
-        
-        viewModel.canRegister
-            .receive(on: DispatchQueue.main)
-            .sink{ [weak self] isEnabled in
-                self?.contentInputCell?.ctaButton.isEnabled = isEnabled
-            }
-            .store(in: &cancellables)
-    }
-}
+        bindTopicSide()
+        bindSideA()
+        bindSideB()
 
-extension TopicGenerateViewController: PHPickerViewControllerDelegate {
-    
-    func picker(_ picker: PHPickerViewController, didFinishPicking results: [PHPickerResult]) {
-
-        picker.dismiss(animated: true) {
-            
-            guard let itemProvider = results.first?.itemProvider else { return }
-            
-            if itemProvider.canLoadObject(ofClass: UIImage.self) {
-                itemProvider.loadObject(ofClass: UIImage.self) { [weak self] image, error in
-                    guard let self = self, let image = image as? UIImage, let option = self.imagePickerOption else { return }
-                    DispatchQueue.main.async {
-                        self.contentInputCell?.setImage(image, option: option)
+        
+        func bindTopicSide() {
+            viewModel.topicSide
+                .receive(on: DispatchQueue.main)
+                .sink{ [weak self] side in
+                    guard let self = self else { return }
+                    switch side {
+                    case .A:
+                        self.aSideView.clear()
+                        self.currentSideView = self.aSideView
+                        self.viewModel.sideAInput(
+                            content: .init(titleDidEnd: self.aSideView.titleSection.contentView.textField.publisher(for: .editingDidEnd), keywordDidEnd: nil),
+                            choices: .init(
+                                choiceA: self.aSideView.optionsSection.contentView.aTextField.textField.anyPublisher(for: .editingDidEnd),
+                                choiceB: self.aSideView.optionsSection.contentView.bTextField.textField.anyPublisher(for: .editingDidEnd))
+                        )
+                    case .B:
+                        self.bSideView.clear()
+                        self.currentSideView = self.bSideView
+                        self.viewModel.sideBInput(
+                            content: .init(
+                                titleDidEnd: self.bSideView.titleSection.contentView.textField.publisher(for: .editingDidEnd),
+                                keywordDidEnd: self.bSideView.keywordSection.contentView.textField.publisher(for: .editingDidEnd)
+                            )
+                        )
                     }
                 }
+                .store(in: &cancellables)
+        }
+        
+        func bindSideA() {
+            
+            viewModel.sideATitleValidation
+                .receive(on: DispatchQueue.main)
+                .sink{ [weak self] (isValid, message) in
+                    if isValid {
+                        self?.aSideView.titleSection.contentView.complete()
+                    }
+                    else {
+                        self?.aSideView.titleSection.contentView.error(message: message)
+                    }
+                }
+                .store(in: &cancellables)
+            
+            viewModel.sideAOptionAValidation
+                .receive(on: DispatchQueue.main)
+                .sink{ [weak self] (isValid, message) in
+                    if isValid {
+                        self?.aSideView.optionsSection.contentView.aTextField.complete()
+                    }
+                    else {
+                        self?.aSideView.optionsSection.contentView.aTextField.error(message: message)
+                        self?.aSideView.optionsSection.contentView.bTextField.errorLabel.text = message
+                    }
+                }
+                .store(in: &cancellables)
+            
+            viewModel.sideAOptionBValidation
+                .receive(on: DispatchQueue.main)
+                .sink{ [weak self] (isValid, message) in
+                    if isValid {
+                        self?.aSideView.optionsSection.contentView.bTextField.complete()
+                    }
+                    else {
+                        self?.aSideView.optionsSection.contentView.bTextField.error(message: message)
+                    }
+                }
+                .store(in: &cancellables)
+            
+            viewModel.sideACanSwitchOption
+                .receive(on: DispatchQueue.main)
+                .sink{ [weak self] isValid in
+                    self?.aSideView.optionSwitch.isEnabled = isValid
+                }
+                .store(in: &cancellables)
+            
+            viewModel.canSideARegister
+                .receive(on: DispatchQueue.main)
+                .sink{ [weak self] canRegister in
+                    self?.aSideView.ctaButton.isEnabled = canRegister
+                }
+                .store(in: &cancellables)
+            
+            viewModel.successRegister = {
+                DispatchQueue.main.async {
+                    self.dismiss(animated: true)
+                }
             }
         }
-    }
-}
+        
+        func bindSideB() {
+            viewModel.sideBTitleValidation
+                .receive(on: DispatchQueue.main)
+                .sink{ [weak self] (isValid, message) in
+                    if isValid {
+                        self?.bSideView.titleSection.contentView.complete()
+                    }
+                    else {
+                        self?.bSideView.titleSection.contentView.error(message: message)
+                    }
+                }
+                .store(in: &cancellables)
+            
+            viewModel.sideBKeywordValidation
+                .receive(on: DispatchQueue.main)
+                .sink{ [weak self] (isValid, message) in
+                    if isValid {
+                        self?.bSideView.keywordSection.contentView.complete()
+                    }
+                    else {
+                        self?.bSideView.keywordSection.contentView.error(message: message)
+                    }
+                }
+                .store(in: &cancellables)
+            
+            viewModel.moveNextInput
+                .throttle(for: .seconds(1), scheduler: DispatchQueue.main, latest: true)
+                .debounce(for: .seconds(1), scheduler: DispatchQueue.main)
+                .sink{ [weak self] _ in
+                    guard let self = self else { return }
+                    self.viewModel.tempStorage = (
+                        title: self.bSideView.titleSection.contentView.textField.text ?? "",
+                        keyword: self.bSideView.keywordSection.contentView.textField.text ?? ""
+                    )
+                    self.coordinator?.startBsideTopicGenerate()
+                }
+                .store(in: &cancellables)
+        }
 
-extension TopicGenerateViewController: UITableViewDelegate, UITableViewDataSource {
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        2
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        
-        switch indexPath.row {
-        case 0:     return inputTableViewCell()
-        case 1:     return contentInputTableViewCell()
-        default:    fatalError()
-        }
-        
-        func inputTableViewCell() -> UITableViewCell {
-            
-            let cell = tableView.dequeueReusableCell(for: indexPath, cellType: TopicInputTableViewCell.self)
-            
-            setGlobalProperty()
-            setLimitCount()
-            setDelegate()
-            setInput()
-            
-            return cell
-            
-            func setGlobalProperty() {
-                inputCell = cell
-            }
-            
-            func setLimitCount() {
-                cell.title.contentView.limitCount = viewModel.limitCount.title
-                cell.keyword.contentView.limitCount = viewModel.limitCount.keyword
-            }
-            
-            func setDelegate() {
-                cell.recommendKeyword.collectionView.delegate = self
-                cell.recommendKeyword.collectionView.dataSource = self
-            }
-            
-            func setInput() {
-                viewModel.input(content: .init(
-                    titleDidEnd: cell.title.contentView.textField.publisher(for: .editingDidEnd),
-                    keywordDidEnd: cell.keyword.contentView.textField.publisher(for: .editingDidEnd)
-                ))
-            }
-        }
-        
-        func contentInputTableViewCell() -> UITableViewCell {
-            
-            let cell = tableView.dequeueReusableCell(for: indexPath, cellType: TopicContentInputTableViewCell.self)
-            
-            setGlobalProperty()
-            setDelegate()
-            setViewModel()
-            
-            func setGlobalProperty() {
-                contentInputCell = cell
-            }
-            
-            func setDelegate() {
-                cell.delegate = self
-            }
-            
-            func setViewModel() {
-                cell.viewModel = viewModel
-            }
-            
-            return cell
-        }
-    }
-}
-
-extension TopicGenerateViewController: TapDelegate {
-    func tap(_ recognizer: DelegateSender) {
-        
-        guard let contentInputCell = contentInputCell else { return }
-        
-        viewModel.register(.init(
-            side: viewModel.topicSide.value,
-            keyword: inputCell?.keyword.contentView.textField.text ?? "",
-            title: inputCell?.title.contentView.textField.text ?? "",
-            choices: [
-                .init(text: contentInputCell.registerText(option: .A), image: contentInputCell.registerImage(option: .A), option: .A),
-                .init(text: contentInputCell.registerText(option: .B), image: contentInputCell.registerImage(option: .B), option: .B)
-            ],
-            deadline: UTCTime.current
-        ))
     }
 }
 
@@ -263,7 +297,6 @@ extension TopicGenerateViewController: UICollectionViewDataSource, UICollectionV
     }
     
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        inputCell?.keyword.contentView.update(text: viewModel.recommendKeywords[indexPath.row])
+        bSideView.keywordSection.contentView.update(text: viewModel.recommendKeywords[indexPath.row])
     }
 }
-*/
